more 10.5 support
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2828 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ ROIDefaultsWindow.m
@@ -33,7 +33,7 @@ static ViewerController		*curController;
 	[[notification object] reloadData];
 }
 
-- (unsigned int)comboBox:(NSComboBox *)aComboBox indexOfItemWithStringValue:(NSString *)aString
+- (NSUInteger)comboBox:(NSComboBox *)aComboBox indexOfItemWithStringValue:(NSString *)aString
 {
 	if( roiNames == 0L) roiNames = [curController generateROINamesArray];
 	
@@ -45,13 +45,13 @@ static ViewerController		*curController;
 	}
 }
 
-- (int)numberOfItemsInComboBox:(NSComboBox *)aComboBox
+- (NSInteger)numberOfItemsInComboBox:(NSComboBox *)aComboBox
 {
 	if( roiNames == 0L) roiNames = [curController generateROINamesArray];
 	return [roiNames count];
 }
 
-- (id)comboBox:(NSComboBox *)aComboBox objectValueForItemAtIndex:(int)index
+- (id)comboBox:(NSComboBox *)aComboBox objectValueForItemAtIndex:(NSInteger)index
 {
     if ( index > -1 )
     {
